+++ rust/operator-binary/src/main.rs
@@ -1,12 +1,12 @@
-use std::sync::Arc;
+use std::{ops::Deref as _, sync::Arc};
 
-use clap::{Parser, crate_description, crate_version};
+use clap::Parser;
 use futures::{StreamExt, pin_mut};
 use hdfs_controller::HDFS_FULL_CONTROLLER_NAME;
 use product_config::ProductConfigManager;
 use stackable_operator::{
     YamlSchema,
-    cli::{Command, ProductOperatorRun},
+    cli::{Command, ProductOperatorRun, RollingPeriod},
     client::{self, Client},
     k8s_openapi::api::{
         apps::v1::StatefulSet,
@@ -27,7 +27,8 @@ use stackable_operator::{
     namespace::WatchNamespace,
     shared::yaml::SerializeOptions,
 };
-use tracing::info_span;
+use stackable_telemetry::{Tracing, tracing::settings::Settings};
+use tracing::{info_span, level_filters::LevelFilter};
 use tracing_futures::Instrument;
 
 use crate::crd::{HdfsCluster, constants::APP_NAME, v1alpha1};
@@ -49,6 +50,9 @@ mod built_info {
 
 pub const OPERATOR_NAME: &str = "hdfs.stackable.tech";
 
+// TODO (@NickLarsenNZ): Change the variable to `CONSOLE_LOG`
+pub const ENV_VAR_CONSOLE_LOG: &str = "HDFS_OPERATOR_LOG";
+
 #[derive(clap::Parser)]
 #[clap(about, author)]
 struct Opts {
@@ -65,23 +69,59 @@ async fn main() -> anyhow::Result<()> {
         Command::Run(ProductOperatorRun {
             product_config,
             watch_namespace,
-            tracing_target,
+            telemetry_arguments,
             cluster_info_opts,
         }) => {
-            stackable_operator::logging::initialize_logging(
-                "HDFS_OPERATOR_LOG",
-                APP_NAME,
-                tracing_target,
+            let _tracing_guard = Tracing::builder()
+                // TODO (@Techassi): This should be a constant
+                .service_name("hbase-operator")
+                .with_console_output((
+                    ENV_VAR_CONSOLE_LOG,
+                    LevelFilter::INFO,
+                    !telemetry_arguments.no_console_output,
+                ))
+                // NOTE (@Techassi): Before stackable-telemetry was used, the log directory was
+                // set via an env: `HDFS_OPERATOR_LOG_DIRECTORY`.
+                // See: https://github.com/stackabletech/operator-rs/blob/f035997fca85a54238c8de895389cc50b4d421e2/crates/stackable-operator/src/logging/mod.rs#L40
+                // Now it will be `ROLLING_LOGS` (or via `--rolling-logs <DIRECTORY>`).
+                .with_file_output(telemetry_arguments.rolling_logs.map(|log_directory| {
+                    let rotation_period = telemetry_arguments
+                        .rolling_logs_period
+                        .unwrap_or(RollingPeriod::Hourly)
+                        .deref()
+                        .clone();
+
+                    Settings::builder()
+                        // TODO (@Techassi): Change to CONSOLE_LOG or FILE_LOG, create constant
+                        .with_environment_variable(ENV_VAR_CONSOLE_LOG)
+                        .with_default_level(LevelFilter::INFO)
+                        .file_log_settings_builder(log_directory, "tracing-rs.log")
+                        .with_rotation_period(rotation_period)
+                        .build()
+                }))
+                .with_otlp_log_exporter((
+                    "OTLP_LOG",
+                    LevelFilter::DEBUG,
+                    telemetry_arguments.otlp_logs,
+                ))
+                .with_otlp_trace_exporter((
+                    "OTLP_TRACE",
+                    LevelFilter::DEBUG,
+                    telemetry_arguments.otlp_traces,
+                ))
+                .build()
+                .init()?;
+
+            tracing::info!(
+                built_info.pkg_version = built_info::PKG_VERSION,
+                built_info.git_version = built_info::GIT_VERSION,
+                built_info.target = built_info::TARGET,
+                built_info.built_time_utc = built_info::BUILT_TIME_UTC,
+                built_info.rustc_version = built_info::RUSTC_VERSION,
+                "Starting {description}",
+                description = built_info::PKG_DESCRIPTION
             );
 
-            stackable_operator::utils::print_startup_string(
-                crate_description!(),
-                crate_version!(),
-                built_info::GIT_VERSION,
-                built_info::TARGET,
-                built_info::BUILT_TIME_UTC,
-                built_info::RUSTC_VERSION,
-            );
             let product_config = product_config.load(&[
                 "deploy/config-spec/properties.yaml",
                 "/etc/stackable/hdfs-operator/config-spec/properties.yaml",